LayoutTests/fast/mediastream/MediaStreamTrack-expected.txt

Tests MediaStreamTrack callbacks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getUserMedia succeeded.
PASS track.readyState is "live"
PASS Track onmute callback succeeded.
PASS track.readyState is "muted"
PASS Track onunmute callback succeeded.
PASS track.readyState is "live"
PASS Track onended callback succeeded.
PASS track.readyState is "ended"
PASS successfullyParsed is true

TEST COMPLETE
